Attempt to implement reverse control with threads
Initial plan was to use two `mpsc::channel`s for communicating,
but sync is needed since output is only valid when engine is blocked waiting
on input. Updated plan is to put a mutex on the output buffer and require
the engine to lock this except when attempting to get input.

Goal of reversed control would be an Arcade object with methods

```
impl Arcade {
    fn new(input: &str) -> Self
    fn get_last_prompt(&self) -> Prompt
    fn push_input(&mut self, &str)
}
```